Move sections selectors out of main redux file
